fix windows not liking ssize_t

+++ src/game/server/gamemodes/zcatch/playerstats.h
@@ -15,7 +15,7 @@ struct CPlayerStats
    private:
     bool m_IsValid;
 
-    ssize_t m_Rank;
+    long m_Rank;
 
    public:
     std::map<std::string, int> m_Data;
@@ -26,8 +26,8 @@ struct CPlayerStats
     CPlayerStats();
     CPlayerStats(int kills, int deaths, int ticksCaught, int ticksIngame, int ticksWarmup, int score, int wins, int fails, int shots);
 
-    void SetRank(ssize_t rank) { m_Rank = rank; };
-    ssize_t GetRank() { return m_Rank;};
+    void SetRank(long rank) { m_Rank = rank; };
+    long GetRank() { return m_Rank;};
 
     CPlayerStats& operator+=(const CPlayerStats& rhs);
